+++ includes/class-wp-document-revisions-recently-revised-widget.php
@@ -251,7 +251,10 @@ class WP_Document_Revisions_Recently_Revised_Widget extends WP_Widget {
 				'wp-i18n',
 			),
 			filemtime( "$dir/$index_js" ),
-			true
+			array(
+				'in_footer' => true,
+				'strategy'  => 'defer',
+			),
 		);
 
 		$index_css = 'css/wpdr-widget-editor-style.css';