+++ includes/class-wp-document-revisions-validate-structure.php
@@ -258,8 +258,11 @@ class WP_Document_Revisions_Validate_Structure {
 	 *
 	 * @param WP_REST_Request $request the arguments to pass to the function.
 	 * @return WP_REST_Response
+	 *
+	 * @global $wpdb Database object.
 	 */
 	public static function correct_document( $request ) {
+		global $wpdb;
 		$wpdr   = self::$parent;
 		$params = $request->get_params();
 		$id     = $params['id'];
@@ -282,7 +285,6 @@ class WP_Document_Revisions_Validate_Structure {
 				}
 				$content = $wpdr->format_doc_id( $parm ) . $content;
 			}
-			global $wpdb;
 			// phpcs:disable WordPress.DB.PreparedSQL.InterpolatedNotPrepared,WordPress.DB.PreparedSQL.NotPrepared,WordPress.DB.DirectDatabaseQuery
 			$post_table = "{$wpdb->prefix}posts";
 			$sql        = $wpdb->prepare(
@@ -311,7 +313,6 @@ class WP_Document_Revisions_Validate_Structure {
 				// replace existing id data.
 				$content = $wpdr->format_doc_id( $parm ) . substr( $content, $end_id + 1 );
 			}
-			global $wpdb;
 			// phpcs:disable WordPress.DB.PreparedSQL.InterpolatedNotPrepared,WordPress.DB.PreparedSQL.NotPrepared,WordPress.DB.DirectDatabaseQuery
 			$post_table = "{$wpdb->prefix}posts";
 			$sql        = $wpdb->prepare(
@@ -357,6 +358,27 @@ class WP_Document_Revisions_Validate_Structure {
 				$name = get_post_meta( $attach_id, '_wp_attached_file', true );
 				update_post_meta( $attach_id, '_wp_attached_file', str_replace( $filename, $new_name, $name ), $name );
 				unlink( $file );
+			}
+
+			// rename attachment post (if no clash).
+			// phpcs:disable WordPress.DB.PreparedSQL.InterpolatedNotPrepared,WordPress.DB.PreparedSQL.NotPrepared,WordPress.DB.DirectDatabaseQuery
+			$post_table = "{$wpdb->prefix}posts";
+			$sql        = $wpdb->prepare(
+				"SELECT COUNT(1) FROM `$post_table` WHERE `post_name` = %s",
+				$new_name
+			);
+			$res        = $wpdb->get_var( $sql );
+			if ( 0 === $res ) {
+				$sql = $wpdb->prepare(
+					"UPDATE `$post_table` SET `post_name` = %s, `post_title` = %s WHERE `id` = %d",
+					$new_name,
+					$new_name,
+					$attach_id
+				);
+				$res = $wpdb->query( $sql );
+				// phpcs:enable WordPress.DB.PreparedSQL.InterpolatedNotPrepared,WordPress.DB.PreparedSQL.NotPrepared,WordPress.DB.DirectDatabaseQuery
+				wp_cache_delete( $id, 'posts' );
+				wp_cache_delete( $id, 'document_revisions' );
 			}
 		}
 
@@ -906,17 +928,25 @@ class WP_Document_Revisions_Validate_Structure {
 			);
 		}
 
-		// check post_title (warning only).
-		$filename = pathinfo( $file, PATHINFO_FILENAME );
-		if ( ! preg_match( '/^[a-f0-9]{32}$/', $filename ) ) {
-			// file does not appear to be md5 encoded.
-			return array(
-				'code'  => 6,
-				'error' => 0,
-				'msg'   => __( 'Document attachment does not appear to be md5 encoded', 'wp-document-revisions' ),
-				'fix'   => 1,
-				'parm'  => $attach_id,
-			);
+		/**
+		 * Filter to Switch off md5 format attachment validation.
+		 *
+		 * @since 3.6
+		 * @param boolean true.
+		 */
+		if ( apply_filters( 'document_validate_md5', true ) ) {
+			// check post_title (warning only).
+			$filename = pathinfo( $file, PATHINFO_FILENAME );
+			if ( ! preg_match( '/^[a-f0-9]{32}$/', $filename ) ) {
+				// file does not appear to be md5 encoded.
+				return array(
+					'code'  => 6,
+					'error' => 0,
+					'msg'   => __( 'Document attachment does not appear to be md5 encoded', 'wp-document-revisions' ),
+					'fix'   => 1,
+					'parm'  => $attach_id,
+				);
+			}
 		}
 
 		return false;