+++ docs/changelog.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### 3.6.0
+
+* NEW: JS scripts will be called with Defer in WP 6.3 onwards.
+* NEW: User pulldowns will show only relevant users.
+* NEW: Filter document_revisions_owner withdrawn as parameter acted on deprecated.
+* NEW: If role already has "read_documents" capability, do not touch on plugin reactivation.
+* NEW: Optionally stop direct access to document files to force access via WordPress.
+* NEW: Filter to switch off attachment MD5 format validation.
+* NEW: Filter document_internal_filename for updating internal file name additionally passed the original name.
+
 ### 3.5.0
 
 " SECURITY: Rest media interface may expose document name.